requirements.txt
streamlit
torch
torchvision
ftfy
regex
tqdm
pillow
requests
git+https://github.com/openai/CLIP.git